requirements.txt
torch>=2.1.0
torchvision
timm>=0.9.0
pyyaml
pycocotools
lvis
pillow
numpy